requirements.txt
# Core dependencies
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# IBM Granite AI models support (Primary AI engine)
ibm-watson-machine-learning>=1.0.312
ibm-cloud-sdk-core>=3.16.0

# PDF processing - Multiple libraries for maximum compatibility
PyPDF2>=3.0.1
pdfplumber>=0.9.0
pymupdf>=1.23.0
pdfminer.six>=20221105
pytesseract>=0.3.10
Pillow>=9.5.0

# Vector storage and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0

# Enhanced text processing
nltk>=3.8.1
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Text processing
tiktoken>=0.5.0
